Extend ActionView::TestCase instead of helper class
